+++ PCIT_core/include/PCIT_core.h
@@ -1,21 +1,14 @@
 //////////////////////////////////////////////////////////////////////
 //                                                                  //
-// Part of the Panthera Project, under the Apache License v2.0      //
+// Part of the PCIT-CPP, under the Apache License v2.0              //
 // You may not use this file except in compliance with the License. //
 // See `http://www.apache.org/licenses/LICENSE-2.0` for info        //
 //                                                                  //
 //////////////////////////////////////////////////////////////////////
 
 
-#include "../include/test.h"
+#pragma once
 
-
-namespace panthera::panther{
-
-
-	auto test() noexcept -> void {
-		evo::printlnCyan("Panther test");
-	};
-
-	
-};
+#include "./UniqueID.h"
+#include "./Diagnostic.h"
+#include "./Printer.h"