+++ PCIT_core/premake5_PCIT_core.lua
@@ -1,0 +1,25 @@
+-- premake5
+
+project "PCIT_core"
+	kind "StaticLib"
+	-- staticruntime "On"
+	
+
+	targetdir(target.lib)
+	objdir(target.obj)
+
+	files {
+		"./src/**.cpp",
+	}
+
+	
+	includedirs{
+		(config.location .. "/libs"),
+	}
+
+	links{
+		"Evo",
+	}
+
+
+project "*"