+++ Panther/include/Panther.h
@@ -1,6 +1,6 @@
 //////////////////////////////////////////////////////////////////////
 //                                                                  //
-// Part of the Panthera Project, under the Apache License v2.0      //
+// Part of the PCIT-CPP, under the Apache License v2.0              //
 // You may not use this file except in compliance with the License. //
 // See `http://www.apache.org/licenses/LICENSE-2.0` for info        //
 //                                                                  //
@@ -9,4 +9,13 @@
 
 #pragma once
 
-#include "./test.h"
+
+#include <PCIT_core.h>
+
+
+#include "./diagnostics.h"
+#include "./Source.h"
+#include "./SourceManager.h"
+#include "./Context.h"
+
+#include "./default_diagnostic_callback.h"