+++ Panther/include/default_diagnostic_callback.h
@@ -1,6 +1,6 @@
 //////////////////////////////////////////////////////////////////////
 //                                                                  //
-// Part of the Panthera Project, under the Apache License v2.0      //
+// Part of the PCIT-CPP, under the Apache License v2.0              //
 // You may not use this file except in compliance with the License. //
 // See `http://www.apache.org/licenses/LICENSE-2.0` for info        //
 //                                                                  //
@@ -8,3 +8,18 @@
 
 
 #pragma once
+
+
+#include <Evo.h>
+#include <PCIT_core.h>
+
+#include "./Context.h"
+
+namespace pcit::panther{
+
+
+	auto createDefaultDiagnosticCallback(const pcit::core::Printer& printer_ref) noexcept 
+		-> Context::DiagnosticCallback;
+
+
+};