+++ Panther/src/SourceManager.cpp
@@ -1,0 +1,84 @@
+//////////////////////////////////////////////////////////////////////
+//                                                                  //
+// Part of the PCIT-CPP, under the Apache License v2.0              //
+// You may not use this file except in compliance with the License. //
+// See `http://www.apache.org/licenses/LICENSE-2.0` for info        //
+//                                                                  //
+//////////////////////////////////////////////////////////////////////
+
+
+#include "../include/SourceManager.h"
+
+namespace pcit::panther{
+
+
+	auto SourceManager::reserveSources(size_t num_sources) noexcept -> void {
+		this->sources.reserve(num_sources);
+	};
+
+
+	// TODO: create custom vector impl to not have to do a push_back
+	// 		(can't do a emplace_back because I want Source::Source() to be private)
+	auto SourceManager::addSource(const std::string& location, const std::string& data) noexcept -> Source::ID {
+		const Source::ID new_source_id = Source::ID(uint32_t(this->sources.size()));
+		this->sources.push_back(Source(new_source_id, location, data));
+		return new_source_id;
+	};
+
+	auto SourceManager::addSource(const std::string& location, std::string&& data) noexcept -> Source::ID {
+		const Source::ID new_source_id = Source::ID(uint32_t(this->sources.size()));
+		this->sources.push_back(Source(new_source_id, location, std::move(data)));
+		return new_source_id;
+	};
+
+	auto SourceManager::addSource(std::string&& location, const std::string& data) noexcept -> Source::ID {
+		const Source::ID new_source_id = Source::ID(uint32_t(this->sources.size()));
+		this->sources.push_back(Source(new_source_id, std::move(location), data));
+		return new_source_id;
+	};
+
+	auto SourceManager::addSource(std::string&& location, std::string&& data) noexcept -> Source::ID {
+		const Source::ID new_source_id = Source::ID(uint32_t(this->sources.size()));
+		this->sources.push_back(Source(new_source_id, std::move(location), std::move(data)));
+		return new_source_id;
+	};
+
+
+	auto SourceManager::addSource(const fs::path& location, const std::string& data) noexcept -> Source::ID {
+		const Source::ID new_source_id = Source::ID(uint32_t(this->sources.size()));
+		this->sources.push_back(Source(new_source_id, location, data));
+		return new_source_id;
+	};
+
+	auto SourceManager::addSource(const fs::path& location, std::string&& data) noexcept -> Source::ID {
+		const Source::ID new_source_id = Source::ID(uint32_t(this->sources.size()));
+		this->sources.push_back(Source(new_source_id, location, std::move(data)));
+		return new_source_id;
+	};
+
+	auto SourceManager::addSource(fs::path&& location, const std::string& data) noexcept -> Source::ID {
+		const Source::ID new_source_id = Source::ID(uint32_t(this->sources.size()));
+		this->sources.push_back(Source(new_source_id, std::move(location), data));
+		return new_source_id;
+	};
+
+	auto SourceManager::addSource(fs::path&& location, std::string&& data) noexcept -> Source::ID {
+		const Source::ID new_source_id = Source::ID(uint32_t(this->sources.size()));
+		this->sources.push_back(Source(new_source_id, std::move(location), std::move(data)));
+		return new_source_id;
+	};
+
+
+
+	
+	auto SourceManager::getSource(Source::ID id) noexcept -> Source& {
+		evo::debugAssert(this->isLocked() == false, "Cannot get a mutable source when the Source Manager is locked");
+		return this->sources[id.get()];
+	};
+
+	auto SourceManager::getSource(Source::ID id) const noexcept -> const Source& {
+		evo::debugAssert(this->isLocked() == false, "Cannot get a mutable source when the Source Manager is locked");
+		return this->sources[id.get()];
+	};
+
+};