+++ libs/Evo.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#if defined(PANTHERA_CONFIG_DEBUG)
+#if defined(PCIT_CONFIG_DEBUG)
 	#define EVO_CONFIG_DEBUG
 #endif