+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+### v0.0.1.0
+Setting up basic frameworks
+#### Panther
+- Added `Context`
+- Added `Source`
+- Added `SourceManager`
+- Added diagnostics
+#### PCIT_core
+- Added `DiagnosticImpl`
+- Added `Printer`
+- Added `UniqueID` and `UniqueComparableID`
+
 ### v0.0.0.0
 - Initial commit
 - Setup build system for Evo, Panther, and pthr